Updated QGIS Plugin
Got the program working in QGIS.
Currently the get_bbox bound only get the bounds, but if a projection that is not epsg:4326 it will cause an issue. There is probably a better way to set up the function, but I just did that to get it working.
Analysis tab layout is set up, but no code controlling it yet.